requirements.txt
# BANCRIPFUTBOT PRO - Dependencias
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.0.3
python-binance==1.0.19
ta==0.10.2
sqlite3
schedule==1.2.0
gunicorn
